requirements.txt
appdirs==1.4.3
basicauth==0.2
boto3==1.4.4
botocore==1.5.70
certifi==2017.4.17
chardet==3.0.3
Django==1.9
django-bootstrap3==8.2.3
django-debug-toolbar==1.8
django-filter==1.0.4
djangorestframework==3.6.3
docutils==0.13.1
future==0.16.0
httplib2==0.10.3
idna==2.5
jmespath==0.9.3
mysqlclient==1.3.10
oauth2==1.9.0.post1
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-twitter==3.3
pytz==2017.2
PyYAML==3.12
requests==2.17.3
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
sqlparse==0.2.3
twitter-ads==1.2.2
urllib3==1.21.1
uWSGI==2.0.15
